Implement --min-element-savings option
When the lead-in or lead-out contain printable characters, it alters the
economics of abbreviating from a screen space standpoint. I don't want
to try to figure out the printable length of lead-in and -out strings
(since they may well have non-printable ANSI escape seqences, for
example), so this options allows for accounting for this friction.

For example, if you want to output an ellipsis for abbreviated path
elements with `-o …`, you could set `-m 2` to ensure there is a net
space savings to abbreviate. There is no sense showing `/sy…` for
`/sys`: it's just less readable and provides no space savings.